+++ src/pages/api/auth/login.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { connectToDatabase } from "@/lib/database";
-import { loginController } from "@/controllers/auth/auth.controller";
+import { loginController } from "@/backend/controllers/auth/auth.controller";
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
 	if (req.method !== "POST") {
