+++ src/pages/api/blogs/slug/[slug].ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { withDatabase } from "@/lib/database";
-import { blogController } from "@/controllers/blog/blog.controller";
+import { blogController } from "@/backend/controllers/blog/blog.controller";
 import { HttpMethod } from "@/helpers/apiHelpers";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
